Minor edit

+++ src/SAMD21turboPWM.h
@@ -6,7 +6,7 @@
 class TurboPWM {
   public:
     void setClockDivider(unsigned int GCLKDiv, bool turbo);
-    int timer(int timernumber, unsigned int TCCDiv, unsigned long int sts, bool fastPWM);
+    int timer(int timernumber, unsigned int TCCDiv, unsigned long long int sts, bool fastPWM);
     int analogWrite(unsigned int pin, unsigned int dutyCycle);  //Duty cycle will be (dutyCycle / _maxDutyCycle) * 100%
     void enable(unsigned int timerNumber, bool enabled);
     float frequency(unsigned int timerNumber);
